requirements.txt
# Awaaz AI - Urdu Voice Generation Platform
# Python dependencies

# Core
python-dotenv==1.0.0
requests==2.31.0

# API Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2

# Audio processing
pydub==0.25.1

# Validation
pydantic[email]==2.5.3
